Create Student Details Management System
This program allows the user to enter details of a set of students, including their name, age, registration number, year of study, course, and grade. The program  then sorts the students in descending order based on their grade and displays the sorted list of students along with their details.

The system also includes a simple password protection feature to prevent unauthorized access. Users are required to enter a password to use the system, and if the password is incorrect, they will not be able to access the system.

This program can be useful for schools or other educational institutions to manage student records, as well as for individual students to keep track of their grades and progress over time. It demonstrates the use of C programming language to manage data and perform sorting algorithms